requirements.txt
av==11.0.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
picamera2==0.3.14
pidng==4.0.9
piexif==1.1.3
Pillow==10.1.0
pyparsing==3.1.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-prctl==1.8.1
simplejpeg==1.7.2
six==1.16.0
v4l2-python3==0.3.2
